misc code clean up

+++ src/vimeo.nim
@@ -314,26 +314,24 @@ proc getVideo(vimeoUrl: string, aId="0", vId="0") =
           echo "<failed to download video stream>"
           includeVideo = false
 
-      if includeAudio:
-        if video.audioStream.exists:
-          reportStreamInfo(video.audioStream)
-          if not grabMulti(video.audioStream.urlSegments, forceFilename=video.audioStream.filename,
-                           forceDl=true).is2xx:
-            echo "<failed to download audio stream>"
-            includeAudio = false
-        else:
+      if includeAudio and video.audioStream.exists:
+        reportStreamInfo(video.audioStream)
+        if not grabMulti(video.audioStream.urlSegments, forceFilename=video.audioStream.filename,
+                         forceDl=true).is2xx:
+          echo "<failed to download audio stream>"
           includeAudio = false
+      else:
+        includeAudio = false
 
       if includeAudio and includeVideo:
         joinStreams(video.videoStream.filename, video.audioStream.filename, safeTitle)
+      elif includeAudio and not includeVideo:
+        convertAudio(video.audioStream.filename, safeTitle, audioFormat)
+      elif includeVideo:
+        moveFile(video.videoStream.filename, finalFilename.changeFileExt(video.videoStream.ext))
+        echo "[complete] ", addFileExt(safeTitle, video.videoStream.ext)
       else:
-        if includeAudio and not includeVideo:
-          convertAudio(video.audioStream.filename, safeTitle, audioFormat)
-        elif includeVideo:
-          moveFile(video.videoStream.filename, finalFilename.changeFileExt(video.videoStream.ext))
-          echo "[complete] ", addFileExt(safeTitle, video.videoStream.ext)
-        else:
-          echo "<no streams were downloaded>"
+        echo "<no streams were downloaded>"
 
 
 proc getProfile(vimeoUrl: string) =
